feat: DiApplication 클래스 추가
